+++ frontend/src/components/SearchForm/index.tsx
@@ -4,19 +4,13 @@ import SearchInput from "@src/components/SearchInput";
 import * as Styled from "./style";
 import useSelectChannels from "@src/hooks/useSelectChannels";
 import useModal from "@src/hooks/useModal";
-import { ChangeEvent, FormEvent, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { PATH_NAME } from "@src/@constants";
-import useSnackbar from "@src/hooks/useSnackbar";
+import useSearchKeywordForm from "@src/hooks/useSearchKeywordForm";
 
 interface Props {
   channelId: number;
 }
 
 function SearchForm({ channelId }: Props) {
-  const navigate = useNavigate();
-  const [searchKeyword, setSearchKeyword] = useState("");
-
   const {
     channelsData,
     channelIds,
@@ -33,40 +27,18 @@ function SearchForm({ channelId }: Props) {
     handleCloseModal: handleCloseSearchOptions,
   } = useModal();
 
-  const { openFailureSnackbar } = useSnackbar();
-
-  const handleSubmit = (event: FormEvent) => {
-    event.preventDefault();
-
-    if (!channelIds.length) {
-      openFailureSnackbar("채널을 하나 이상 선택 후 검색 버튼을 눌러주세요.");
-
-      return;
-    }
-
-    if (!searchKeyword.length) {
-      openFailureSnackbar(
-        "검색할 키워드를 입력하신 후 검색 버튼을 눌러주세요."
-      );
-
-      return;
-    }
-
-    navigate(
-      `${PATH_NAME.SEARCH_RESULT}?keyword=${searchKeyword}&channelIds=${channelIds}`
-    );
-  };
-
-  const handleChangeSearchKeyword = (event: ChangeEvent<HTMLInputElement>) => {
-    setSearchKeyword(event.target.value);
-  };
+  const {
+    searchKeyword,
+    handleChangeSearchKeyword,
+    handleSubmitSearchKeyword,
+  } = useSearchKeywordForm({ channelIds });
 
   return (
     <>
       {isSearchInputFocused && (
         <Dimmer hasBackgroundColor={false} onClick={handleCloseSearchOptions} />
       )}
-      <Styled.Container onSubmit={handleSubmit}>
+      <Styled.Container onSubmit={handleSubmitSearchKeyword}>
         <SearchInput
           placeholder="검색 할 키워드를 입력해주세요."
           onFocus={handleOpenSearchOptions}
